[mono] Move the jit info table definitions into a separate jit-info.h file, it no longer depends on domains.
